apps/ai-service/requirements.txt
fastapi
uvicorn
pydantic
requests
redis
google-generativeai